edited implementation of createData

+++ src/main/java/com/kubraboduroglu/projects/earthquakes/service/UsgsClientService.java
@@ -1,15 +1,8 @@
 package com.kubraboduroglu.projects.earthquakes.service;
 
-import java.util.List;
-
 import org.springframework.http.ResponseEntity;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonMappingException;
-import com.kubraboduroglu.projects.earthquakes.dto.UsgsQueryReqDTO;
 import com.kubraboduroglu.projects.earthquakes.dto.UsgsResponseDTO;
-
-import reactor.core.publisher.Mono;
 
 public interface UsgsClientService {
 
@@ -17,5 +10,5 @@ public interface UsgsClientService {
 	ResponseEntity<String> getCountEntity(String startTime, String endTime);
 	ResponseEntity<String> getCountWithWebClient(String startTime, String endTime);
 	ResponseEntity<String> getUsgsData(String startTime, String endTime, Double minMagnitude);
-	UsgsResponseDTO getUsgsDatav2(String startTime, String endTime, Double minMagnitude);
+	UsgsResponseDTO getUsgsDatav2(String startTime, String endTime, Integer minMagnitude);
 }
